+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+<a name="0.5.0"></a>
+# [0.5.0](https://github.com/termage/termage) (2021-10-07)
+* Added new `getTerminal` method for Termage and Element class.
+
+### BREAKING CHANGES
+
+* method `renderer` renamed to `output`
+
 <a name="0.4.0"></a>
 # [0.4.0](https://github.com/termage/termage) (2021-10-06)
 * Updated colors for default theme.